[#1]feat: implement configure method in cells
