+++ ProjectManager/ProjectManager/Cell/CellConfigurable.swift
@@ -1,0 +1,12 @@
+//
+//  CellConfigurable.swift
+//  ProjectManager
+//
+//  Created by Seungjin Baek on 2021/07/01.
+//
+
+import UIKit
+
+protocol CellConfigurable {
+    func configure(tasks: [Task], rowAt row: Int)
+}
