--- ProjectManager/ProjectManager/DoingTableViewCell.swift
@@ -1,27 +1,0 @@
-//
-//  DoingTableViewCell.swift
-//  ProjectManager
-//
-//  Created by 이성노 on 2021/06/29.
-//
-
-import UIKit
-
-class DoingTableViewCell: UITableViewCell {
-
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var contentLabel: UILabel!
-    @IBOutlet weak var dateLabel: UILabel!
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-}
