--- ProjectManager/ProjectManager/ToDoTableViewCell.swift
@@ -1,29 +1,0 @@
-//
-//  ToDoTableViewCell.swift
-//  ProjectManager
-//
-//  Created by James, Jay, Ian on 2021/06/29.
-//
-
-import UIKit
-
-class ToDoTableViewCell: UITableViewCell {
-
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var contentLabel: UILabel!
-    @IBOutlet weak var dateLabel: UILabel!
-    
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        // Initialization code
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-        // Configure the view for the selected state
-    }
-
-    
-}
